Fixes paths with spaces
As it is, when trying to call lazygit with a file that contains a space char in its path will fail, as the path will be broken into two arguments when passed to lazygit. The quotes I added fixes that. Then, there is the fact that the path returned by git has a newline at the end, so I trim the path to remove newlines at the beginning and end of paths. This version works with projects I have inside "Google Drive".

I DID NOT TEST THIS, but are basically the same changes I made on the other branch.