Don't use -p if GIT_DIR or GIT_WORK_TREE are set
Setting `-p <path>` is the same as setting `-w <path> -g <path>/.git/`,
but this eliminates the possibility that the git dir could be outside of
the work tree, or else named something besides `.git`. Use case for me
is my configuration files are VC'ed in a git repo, and checked out into
`$HOME` (the work tree). The repo is a bare repo elsewhere.